only:refactoring: Now `TestBase::updateRpm` method accepts `float` instead of `int`

+++ unit_tests/tests/util/test_base.cpp
@@ -23,7 +23,7 @@ void TestBase::periodicSlowCallback() {
     engine->periodicSlowCallback();
 }
 
-void TestBase::updateRpm(const int rpm) {
+void TestBase::updateRpm(const float rpm) {
     Sensor::setMockValue(SensorType::Rpm, rpm);
     periodicFastCallback();
 }
